fix(installations): set hard-coded per_page
GitHub's installations end-point only returns the first 30 repositories
belonging to an installation, even if there are more repositories
belonging to that installation.

Set hard-coded `per_page` value when fetching repositories belonging
to an installation. Value is set to 100 because that is the maximum that
GitHub will return for any single API fetch.

This is a temporary change because some installations may have more
than 100 repositories belonging to the installation.